export testing suite for bucket storage, review comment changes
